Initial commit.
So far I've got the board view with an array of nodes rendering
correctly. There's css and view rules for creating tokens, but
they don't seem to be rendering correctly into the final view.

Resizing seems to work upwards, although it isn't perfect. Resizing
down looks pretty awful. I'm convinced this is because of the borders
on the nodes, since it seems to work without that. Margin -1px helps,
but doesn't alleviate all of the problems.