Drag n' Drop
Still total prototype stage. Drag and drop onto the board works. Not terribly
clean because there are references abound to make the events work. You can drag
on the box o' pieces and dropping onto the board makes the event work.

You can't currently drag on the pieces to move them, so that kinda blows. Just
making progress in the direction of having a fully functional prototype.