Game board is aware of chains
This means that if you surround a chain of pieces that is continuous
with your own pieces, it will be captured. This could be vastly
improved with fewer references to other objects if I used the observer
pattern and used events to pass around information like "hey this chain
just went away, clean up your reference" instead of just keeping a reference
around forever. Good enough for a simple demo though.