Circular references are bad and cause bugs
I knew there'd be some hilarious bug involved with having nodes refer back
to their chain. Because of how I wrote the logic to merge chains together, we
only ever merged the ORIGINAL chain for each node instead of the entire chain
as it was built up. In practice, you could only ever merge up to 5 nodes at
a time with up to 3 other nodes, because of how we'd keep track of an original
chain.

The fix for this was a one liner (update the chain reference when you merge nodes
from one chain to another) but in general, the way to prevent this from happening
is to NOT have circular references from children elements to their parent elements
and instead use OO and event-listener / delegate patterns.